create geometry and advanced_control packages. split geometry into deck and util. move DeckItem into geometry.
