+++ api/src/opentrons/api/session.py
@@ -22,7 +22,7 @@ from opentrons.types import Location, Point
 from opentrons.calibration_storage import helpers
 from opentrons.protocol_api import (ProtocolContext,
                                     labware)
-from opentrons.protocols.api_support import module_geometry
+from opentrons.protocols.geometry import module_geometry
 from opentrons.protocols.execution.execute import run_protocol
 from opentrons.hardware_control import (API, ThreadManager,
                                         ExecutionCancelledError,
