+++ api/src/opentrons/commands/commands.py
@@ -11,7 +11,7 @@ from opentrons.legacy_api.containers import (Well as OldWell,
                                              Slot as OldSlot,
                                              location_to_list)
 from opentrons.protocol_api.labware import Well, Labware
-from opentrons.protocols.api_support.module_geometry import ModuleGeometry
+from opentrons.protocols.geometry.module_geometry import ModuleGeometry
 from opentrons.protocols.api_support.util import FlowRates
 from opentrons.types import Location
 from opentrons.drivers import utils
