+++ api/src/opentrons/protocol_api/labware.py
@@ -29,11 +29,12 @@ from opentrons.types import Location, Point
 from opentrons.protocols.types import APIVersion
 from opentrons_shared_data import load_shared_data, get_shared_data_root
 from opentrons.protocols.api_support.definitions import (
-    MAX_SUPPORTED_VERSION, DeckItem)
+    MAX_SUPPORTED_VERSION)
+from opentrons.protocols.geometry.deck_item import DeckItem
 from opentrons.protocols.api_support.constants import (
     OPENTRONS_NAMESPACE, CUSTOM_NAMESPACE, STANDARD_DEFS_PATH, USER_DEFS_PATH)
 if TYPE_CHECKING:
-    from opentrons.protocols.api_support.module_geometry import ModuleGeometry  # noqa(F401)
+    from opentrons.protocols.geometry.module_geometry import ModuleGeometry  # noqa(F401)
     from opentrons_shared_data.labware.dev_types import (
         LabwareDefinition, LabwareParameters, WellDefinition)
