+++ api/tests/opentrons/protocol_api/test_context.py
@@ -5,6 +5,7 @@ from unittest import mock
 
 import opentrons.protocol_api as papi
 import opentrons.protocols.api_support as papi_support
+import opentrons.protocols.geometry as papi_geometry
 from opentrons_shared_data import load_shared_data
 from opentrons_shared_data.pipette import name_for_model
 from opentrons.types import Mount, Point, Location, TransferTipPolicy
@@ -13,7 +14,7 @@ from opentrons.hardware_control.pipette import Pipette
 from opentrons.hardware_control.types import Axis
 from opentrons.config.pipette_config import config_models
 from opentrons.protocol_api import paired_instrument_context as paired
-from opentrons.protocols.api_support import transfers as tf
+from opentrons.protocols.advanced_control import transfers as tf
 from opentrons.protocols.types import APIVersion
 from opentrons.calibration_storage import get, types as cs_types
 
@@ -133,7 +134,7 @@ async def test_location_cache(loop, monkeypatch, get_labware_def, hardware):
                 (Point(1, 2, 10), None),
                 (Point(1, 2, 3), None)]
 
-    monkeypatch.setattr(papi_support.geometry, 'plan_moves', fake_plan_move)
+    monkeypatch.setattr(papi_geometry.util, 'plan_moves', fake_plan_move)
     # When we move without a cache, the from location should be the gantry
     # position
     right.move_to(lw.wells()[0].top())
